requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
requests==2.31.0
fastapi==0.94.1
uvicorn==0.22.0
sqlalchemy==2.0.20
alembic==1.11.1
httpx==0.24.1
pydantic==1.10.11
pytest==7.4.0
